Fix pr crossing for sysvar keyed-accounts
